# Rework communication protocol