--- a/client.conf
+++ b/client.conf
@@ -1,6 +1,6 @@
 {
-    "verbosity": "debug",
+    "verbosity": "info",
     "pid_file": "/var/run/memflow.pid",
     "log_file": "/var/log/memflow.log",
-    "socket_addr": "tcp://127.0.0.1:8000"
+    "socket_addr": "http://127.0.0.1:8000"
 }