--- a/daemon.conf
+++ b/daemon.conf
@@ -1,6 +1,6 @@
 {
-    "verbosity": "debug",
+    "verbosity": "info",
     "pid_file": "/var/run/memflow.pid",
     "log_file": "/var/log/memflow.log",
-    "socket_addr": "unix:/var/run/memflow.sock"
+    "socket_addr": "127.0.0.1:8000"
 }