--- a/memflow-cli/src/commands/benchmark.rs
+++ b/memflow-cli/src/commands/benchmark.rs
@@ -1,9 +1,17 @@
 use crate::Config;
-use memflow_client::dispatch::benchmark;
 
 use clap::{App, Arg, ArgMatches, SubCommand};
 
-use log::trace;
+use log::{error, trace};
+use memflow_client::dispatch::{
+    create_client, create_client_async, dispatch_request, dispatch_request_async_client,
+    dispatch_request_client,
+};
+
+use memflow_daemon::memflow_rpc::{
+    ProcessInfoRequest, ReadPhysicalMemoryEntryRequest, ReadPhysicalMemoryRequest,
+    ReadVirtualMemoryEntryRequest, ReadVirtualMemoryRequest,
+};
 
 pub const COMMAND_STR: &str = "benchmark";
 
@@ -70,4 +78,166 @@ pub fn handle_command(conf: &Config, matches: &ArgMatches) {
         .expect("bool parse failed, pysical must be true or false");
 
     benchmark(conf, physical_mode, conn_id, read_size, async_mode)
+}
+
+pub fn benchmark(
+    conf: &Config,
+    physical_mode: bool,
+    conn_id: &str,
+    read_size: u64,
+    async_mode: bool,
+) {
+    if async_mode {
+        benchmark_async(conf, physical_mode, conn_id, read_size);
+    } else {
+        benchmark_sync(conf, physical_mode, conn_id, read_size);
+    }
+}
+
+fn benchmark_sync(conf: &Config, physical_mode: bool, conn_id: &str, read_size: u64) {
+    let pid = 0;
+    let address = dispatch_request(
+        conf,
+        ProcessInfoRequest {
+            conn_id: conn_id.to_string(),
+            pid: pid,
+        },
+    )
+    .expect("could not access process info")
+    .process
+    .unwrap()
+    .address;
+
+    let entry = ReadVirtualMemoryEntryRequest {
+        addr: address,
+        len: read_size,
+    };
+    let req = ReadVirtualMemoryRequest {
+        conn_id: conn_id.to_string(),
+        pid: pid,
+        base_offsets: false,
+        reads: vec![entry],
+    };
+    let phys_entry = ReadPhysicalMemoryEntryRequest {
+        addr: address,
+        len: read_size,
+    };
+    let phys_req = ReadPhysicalMemoryRequest {
+        conn_id: conn_id.to_string(),
+        reads: vec![phys_entry],
+    };
+
+    let (mut client, rt) = create_client(conf);
+
+    let start_time = std::time::Instant::now();
+    let mut total_runs = 0;
+    loop {
+        total_runs += 1;
+
+        let response = if !physical_mode {
+            dispatch_request_client(conf, req.clone(), &mut client, &rt).map(|_| ())
+        } else {
+            dispatch_request_client(conf, phys_req.clone(), &mut client, &rt).map(|_| ())
+        };
+        match response {
+            Err(e) => error!("{:#?}", e),
+            Ok(_) => (),
+        }
+
+        if (std::time::Instant::now() - start_time).as_secs() > 10 {
+            break;
+        }
+    }
+    let end_time = std::time::Instant::now();
+
+    let total_sec = (end_time - start_time).as_secs_f64();
+    println!(
+        "Total: {} s, Total: {}, Each: {} ms",
+        total_sec,
+        total_runs,
+        total_sec * 1000.0 / total_runs as f64
+    );
+}
+
+fn benchmark_async(conf: &Config, physical_mode: bool, conn_id: &str, read_size: u64) {
+    let pid = 0;
+    let address = dispatch_request(
+        conf,
+        ProcessInfoRequest {
+            conn_id: conn_id.to_string(),
+            pid: pid,
+        },
+    )
+    .expect("could not access process info")
+    .process
+    .unwrap()
+    .address;
+
+    let entry = ReadVirtualMemoryEntryRequest {
+        addr: address,
+        len: read_size,
+    };
+    let req = ReadVirtualMemoryRequest {
+        conn_id: conn_id.to_string(),
+        pid: pid,
+        base_offsets: false,
+        reads: vec![entry],
+    };
+    let phys_entry = ReadPhysicalMemoryEntryRequest {
+        addr: address,
+        len: read_size,
+    };
+    let phys_req = ReadPhysicalMemoryRequest {
+        conn_id: conn_id.to_string(),
+        reads: vec![phys_entry],
+    };
+
+    let rt = tokio::runtime::Runtime::new().unwrap();
+
+    let start_time = std::time::Instant::now();
+    let mut total_runs = 0;
+
+    let bench = async {
+        let client = create_client_async(conf).await;
+        let mut responses = vec![];
+        loop {
+            total_runs += 1;
+
+            let response = async {
+                if !physical_mode {
+                    let mut client_cp = client.clone();
+                    dispatch_request_async_client(conf, req.clone(), &mut client_cp)
+                        .await
+                        .map(|_| ())
+                } else {
+                    let mut client_cp = client.clone();
+                    dispatch_request_async_client(conf, phys_req.clone(), &mut client_cp)
+                        .await
+                        .map(|_| ())
+                }
+            };
+            responses.push(response);
+
+            if (std::time::Instant::now() - start_time).as_secs() > 10 || total_runs >= 20000 {
+                break;
+            }
+        }
+        let results = futures::future::join_all(responses).await;
+        for res in results {
+            match res {
+                Err(e) => error!("{:#?}", e),
+                Ok(_) => (),
+            }
+        }
+    };
+    rt.block_on(bench);
+    let end_time = std::time::Instant::now();
+
+    let total_sec = (end_time - start_time).as_secs_f64();
+    println!(
+        "Total: {} s, Total: {}, Each: {} ms",
+        total_sec,
+        total_runs,
+        total_sec * 1000.0 / total_runs as f64
+    );
 }