--- a/memflow-cli/src/commands/gdb/ls.rs
+++ b/memflow-cli/src/commands/gdb/ls.rs
@@ -1,20 +1,25 @@
-use crate::dispatch::*;
 use crate::Config;
 
 use clap::{App, ArgMatches, SubCommand};
 
-use log::trace;
+use log::{error, trace};
 
-use memflow_daemon::request;
+use memflow_client::dispatch::dispatch_request;
+use memflow_daemon::memflow_rpc::GdbListRequest;
 
 pub const COMMAND_STR: &str = "ls";
 
 pub fn command_definition<'a, 'b>() -> App<'a, 'b> {
-    SubCommand::with_name(COMMAND_STR).about("lists all mounted file systems")
+    SubCommand::with_name(COMMAND_STR).about("lists all gdb stubs")
 }
 
 pub fn handle_command(conf: &Config, _matches: &ArgMatches) {
     trace!("handling command");
 
-    dispatch_request(conf, request::Message::GdbList).unwrap();
+    let result = dispatch_request(conf, GdbListRequest {});
+
+    match result {
+        Err(e) => error!("{:#?}", e),
+        Ok(r) => println!("{:#?}", r.stubs),
+    }
 }