--- a/memflow-cli/src/commands/mod.rs
+++ b/memflow-cli/src/commands/mod.rs
@@ -1,8 +1,6 @@
+pub mod benchmark;
 pub mod connection;
 pub mod proc;
-pub mod benchmark;
 
-// TODO
-//pub mod fuse;
-
-//pub mod gdb;
+pub mod fuse;
+pub mod gdb;