--- a/memflow-cli/src/commands/proc/ls.rs
+++ b/memflow-cli/src/commands/proc/ls.rs
@@ -1,11 +1,9 @@
-use crate::dispatch::*;
+use memflow_client::dispatch::dispatch_request;
 use crate::Config;
 
 use clap::{App, Arg, ArgMatches, SubCommand};
 
-use log::trace;
-
-use memflow_daemon::request;
+use log::{error, trace};
 
 pub const COMMAND_STR: &str = "ls";
 
@@ -13,10 +11,10 @@ const CONNECTION_ID: &str = "CONNECTION_ID";
 
 pub fn command_definition<'a, 'b>() -> App<'a, 'b> {
     SubCommand::with_name(COMMAND_STR)
-        .about("lists all open connections")
+        .about("lists all processes")
         .arg(
             Arg::with_name(CONNECTION_ID)
-                .help("the connector to be used for the new connection")
+                .help("the connector to be used for process listing")
                 .index(1)
                 .required(true),
         )
@@ -27,11 +25,19 @@ pub fn handle_command(conf: &Config, matches: &ArgMatches) {
 
     let conn_id = matches.value_of(CONNECTION_ID).unwrap();
 
-    dispatch_request(
+    let result = dispatch_request(
         conf,
-        request::Message::ListProcesses(request::ListProcesses {
+        memflow_daemon::memflow_rpc::ListProcessesRequest {
             conn_id: conn_id.to_string(),
-        }),
-    )
-    .unwrap();
+        },
+    );
+    
+    match result {
+        Err(e) => error!("{:#?}", e),
+        // Ok(r) => println!("{:#?}", "asdf"),
+        Ok(r) => {
+            let s: Vec<String> = r.processes.iter().map(|x| format!("Name: {}, Pid: {}", x.name, x.pid)).collect();
+            println!("{:#?}", s)
+        },
+    }
 }