--- a/memflow-client/src/lib.rs
+++ b/memflow-client/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod dispatch;