--- a/memflow-daemon/build.rs
+++ b/memflow-daemon/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    tonic_build::compile_protos("../memflow_rpc.proto").unwrap();
+}