This implements the communication protocol between cli/connector and the daemon with the cross-platform and cross language gRPC standard using tonic.

The features I added:
- tonic rpc
- New daemon commands: virt_mem read/write, process_info
- New cli commands: process_info, benchmark async/sync requests
- New connector example/benchmark: read_virt

What is not implemented yet and further improvements:
- gdb commands
- fuse commands
- memflow_daemon virt_mem command: speedup with memflow_daemon::state::CachedWin32Process

Connector benchmarks perform 2.8x (virtual) and 4x (physical) faster:
```
$ ping <host>
rtt min/avg/max/mdev = 0.365/0.473/0.609/0.068 ms

Before:
$ cargo run --release --example=read_virt -- --host tcp://<host>:8000 --id <id>
[read_virt] 724.952878062926 reads/sec
[read_virt] 1.3794 ms/read
[read_virt] 720.4610951008646 reads/sec
[read_virt] 1.388 ms/read

$ cargo run --release --example=read_phys -- --host tcp://<host>:8000 --id <id>
[read_phys] 559.1902924565229 reads/sec
[read_phys] 1.7883 ms/read
[read_phys] 592.6101514118936 reads/sec
[read_phys] 1.68745 ms/read

After:
$ cargo run --release --example=read_virt -- --host http://<host>:8000 --id <id>
[read_virt] 2063.983488132095 reads/sec
[read_virt] 0.4845 ms/read
[read_virt] 2050.8613617719443 reads/sec
[read_virt] 0.4876 ms/read

$ cargo run --release --example=read_phys -- --host http://<host>:8000 --id <id>
[read_phys] 2239.6416573348265 reads/sec
[read_phys] 0.4465 ms/read
[read_phys] 2372.479240806643 reads/sec
[read_phys] 0.4215 ms/read
```

Other benchmarks using cli. Sync is comparable with the above benchmarks while async shows the performance of concurrent execution of multiple requests at the same time.
```
sync, physical:
$ cargo run --release --bin memflow-cli -- --config client.conf benchmark 0838a91f90 64 false true
Total: 11.0000146 s, Total: 32080, Each: 0.34289322319202 ms
sync, virtual:
$ cargo run --release --bin memflow-cli -- --config client.conf benchmark 0838a91f90 64 false false
Total: 11.001185069 s, Total: 2956, Each: 3.721645828484438 ms
async, physical:
$ cargo run --release --bin memflow-cli -- --config client.conf benchmark 0838a91f90 64 true true
Total: 0.99157288 s, Total: 20000, Each: 0.049578644 ms
async, virtual:
$ cargo run --release --bin memflow-cli -- --config client.conf benchmark 0838a91f90 64 true false
Total: 53.264028407 s, Total: 20000, Each: 2.66320142035 ms
```
Virtual memory access is much slower because CachedWin32Process is not yet used. Once this is implemented, virtual access should be at the same performance level as physical access.