@ko1N 2021-02-07
Thanks for this awesome pull request! This also sounds like a great step moving forward towards multi-daemon support ("memflow-cloud").

Luckily we are reworking the memflow os layer abstraction atm for the 0.2 release so it should simplify everything regarding Win32Kernel / Process.