@ko1N 2021-02-10
I will take a closer look at the PR on the weekend - we might not be able to merge it yet as it's missing some core features. It'd be a good candidate to merge it on the `next` branch however for the 0.2 release. It would be nice if you can add the missing commands but if not I will take care of that.

As a side note please make sure to properly read and understand the contribution guidelines you can find here: https://github.com/memflow/memflow-cli/blob/master/CONTRIBUTE.md

Thanks again for the work on this PR.