@youduda 2021-02-11
I fixed the code styling and added the missing gdb and fuse commands.

Well, I'm curious, what the "memflow-cloud" is all about. One client instance being able to connect to multiple daemons at the same time differentiating with e.g. the connection id? If yes, then the dispatch_request_[async]_client() interfaces indeed provide what is required for this.

For my project I need a way to access a chain of dependent virtual memory locations (e.g. the 10th element of a linked list) in a process with low latency. Because each request depends on the previous response the network latency sums up quite a lot. To solve this I plan to add a cache on top of memflow-client automatically prefetching memory regions that have been used previously. The code should be short because (of course) there are great crates doing most of the work: dyn_cache looks promising and maybe moka.
Because memflow is capable of such a huge amount of things including caching, I wonder if there is already something that provides what I need.