@ko1N 2021-02-11
You're right. The goal is to let one CLI client connect to multiple daemons at the same time (and also load/execute plugins/scripts on the daemons directly). Plugins is something we want to add for the next iteration of the memflow-daemon. This should also be much faster than working over the daemon-connector itself.

You're right regarding caches. Memflow has something similar to a TLB cache and a physical page cache (which caches physical pages) by default. It might be worth increasing the cache sizes over the wire as the bottleneck becomes larger. For local usage, the cache size should correlate to the actual CPU cache size. This prevents slowdowns due to too many cache misses. However, for high-latency applications, it might be worth increasing the cache size drastically.

If you use the daemon connector as a physical-memory provider the caching should be enabled on the consuming application. If you bake in virtual memory access into the daemon itself (which reduces the amount of physical reads quite a bit) then the cache can be moved into the daemon connector / cli client itself. The new os-layers plugin branch on the memflow repo should also make it much easier to streamline this as you do not need to have typed Win32 objects anymore. The OS is abstracted in a plugin and can be used generically like a connector.